change back endpoint accept message rule
If more than one GCSes connected, there are cases that flight control
forwards message from one GCS to other GCS, and the forwarded message
keeps the source sysid 255 which is the same as GCS. In this case,
255/x would be recorded as known sysid/compid in flightcontrol endpoint,
and then GCS messages with source of 255/x would be wrongly rejected.
So we need to change back the rule to check whether the endpoint has
other sysid/compid before reject message with known source sysid/compid.

Change-Id: I36ac3d35d5f95e63843e83e069d293cec14cc5b7
Signed-off-by: mazhiqiang <mazhiqiang@fishsemi.com>